Configure vitest
Setting 'globals: true' works fine on the CLI but VSCode still has a fit
that it can't find 'test' and 'expect'. Importing manually seems to
work, but it's an annoying workaround.

VSCode also complains about the triple-slash declaration in
vite.config.ts and says that the file isn't part of the project. I think
it doesn't recognize the "references" entry for tsconfig.node.json that
lists vite.config.ts. Having to ignore the config file is annoying, but
it's a small file so the risk is not so bad I suppose.

But holy crap the js/ts community needs to get its shit together.